revert to more basic allocation strategy

+++ Source/common.h
@@ -49,7 +49,7 @@ static inline float scale_linear(float x, float min, float max, float newmin, fl
     return (x - min) / (max - min) * (newmax - newmin) + newmin;
 }
 
-static inline u64 nextPowTwo(u64 n) {
+static inline u32 nextPowTwo(u32 n) {
     n--;
     n |= n >> 1;
     n |= n >> 2;
